feat #1 이미지 보여주기 및 컴포넌트화
