+++ Study/Ham/Week2/package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "react": "18.2.0",
     "react-native": "0.71.10",
+    "react-native-fix-image": "^2.1.0",
     "react-native-image-crop-picker": "^0.40.0",
     "react-native-image-resizer": "^1.4.5",
     "react-native-permissions": "^3.8.0",
